docker/data-loader/requirements.txt
pandas==2.0.3
pyarrow==12.0.1
psycopg2-binary==2.9.7
numpy==1.24.3
sqlalchemy==2.0.19
